Move "git repo" wrapper into its own class
also add minimal tests for the methods I will be changing later
